refactor: 💡 move watering schedules to separate folder
Move schedules to own folder in order to increase maintainability.
