+++ server/data-store/src/schema/index.ts
@@ -2,11 +2,7 @@ import {
     GraphQLSchema,
     GraphQLObjectType,
     GraphQLString,
-    GraphQLInputObjectType,
     GraphQLNonNull,
-    GraphQLInt,
-    GraphQLList,
-    GraphQLBoolean,
 } from 'graphql'
 
 import {
@@ -18,21 +14,23 @@ import {
     verifyAndDecodeToken,
 } from '../lib/auth'
 
-import {
-    getWateringSchedulesForUser,
-    getWateringScheduleById,
-    scheduleWateringFor,
-    removeWateringScheduleById,
-    updateWateringSchedule
-} from '../lib/wateringSchedules'
 import * as log from '../logging'
 import {
     parseTokenFromHeaders,
     parseUserIdFromToken
 } from '../lib/http'
 
+import {
+    wateringScheduleById,
+    wateringSchedulesByUserId,
+    nextWateringDateFor,
+    updateWateringScheduleForUser,
+    deleteWateringSchedule
+} from '../wateringSchedules'
+
+import { Status } from './types'
+
 const nonNullGqlString = { type: new GraphQLNonNull(GraphQLString) }
-const interval = { type: GraphQLInt, description: "The interval (in days) with which schedules should be updated where interval >= 1." }
 
 /**
  * Types
@@ -59,41 +57,6 @@ const Registration = new GraphQLObjectType({
     })
 })
 
-const WateringSchedule = new GraphQLObjectType({
-    name: 'WateringSchedule',
-    description: 'Information about a plant and when to water next.',
-    fields: () => ({
-        id: nonNullGqlString,
-        userId: nonNullGqlString,
-        plants: { type: GraphQLList(Plant) },
-        nextTimeToWater: nonNullGqlString,
-        interval
-    })
-})
-const plantInput = new GraphQLInputObjectType({
-    name: 'plantInput',
-    fields: {
-        id: { type: GraphQLString, description: 'uuid of the Plant' },
-        name: nonNullGqlString,
-    }
-})
-
-const Plant = new GraphQLObjectType({
-    name: "Plant",
-    description: "Information about a plant.",
-    fields: () => ({
-        id: { type: GraphQLString, description: 'uuid of the Plant' },
-        name: nonNullGqlString,
-    })
-})
-const Status = new GraphQLObjectType({
-    name: "Status",
-    description: "Status of operation, evaluates to True or False",
-    fields: () => ({
-        status: { type: GraphQLBoolean }
-    })
-})
-
 /**
  * Queries
  */
@@ -101,52 +64,8 @@ const queryType = new GraphQLObjectType({
     name: 'Query',
     description: 'Query operations',
     fields: () => ({
-        wateringScheduleById: {
-            name: 'wateringScheduleById',
-            description: 'The watering schedule by schedule id',
-            type: WateringSchedule,
-            args: {
-                id: nonNullGqlString,
-            },
-            resolve: async (_root, args, context) => {
-                return parseTokenFromHeaders(context)
-                    .then(verifyAndDecodeToken)
-                    .then(parseUserIdFromToken)
-                    .then(userId => {
-                        log.info('Retrieving schedule for user: ', userId)
-                        return getWateringScheduleById(args.id)
-                    })
-                    .catch(err => { log.error(err); throw new Error('Invalid Request') })
-            }
-        },
-        wateringScheduleForUser: {
-            name: 'wateringScheduleForUser',
-            description: 'The watering schedules for a user by user id',
-            type: GraphQLList(WateringSchedule),
-            args: {
-                offset: {
-                    type: GraphQLString,
-                    description: 'The document id to start from.'
-                },
-                limit: {
-                    type: GraphQLInt,
-                    description: 'The number of schedules to retrieve. Defaults to 10'
-                }
-            },
-            resolve: async (_root, args, context) => {
-                const limit = args.limit ?? 10
-                return parseTokenFromHeaders(context)
-                    .then(verifyAndDecodeToken)
-                    .then(parseUserIdFromToken)
-                    .then(userId => {
-                        log.info('Retrieve watering schedules for ', userId)
-                        log.debug('with offset:', args.offset, 'and limit:', limit)
-                        return getWateringSchedulesForUser(userId, args.offset ? args.offset : null, limit)
-                    }
-                        )
-                    .catch(err => { log.error(err); throw new Error('Invalid Request') })
-            }
-        },
+        wateringScheduleById,
+        wateringSchedulesByUserId
     })
 })
 
@@ -215,66 +134,9 @@ const mutationType = new GraphQLObjectType({
                     .catch(err => { log.error(err); throw new Error('Invalid Request') })
             }
         },
-        nextWateringDateFor: {
-            description: "Set up a reminder for when to water a plant next.",
-            type: WateringSchedule,
-            args: {
-                plants: {
-                    type: GraphQLList(plantInput)
-                },
-                timestamp: nonNullGqlString,
-                interval
-            },
-            resolve: async (_root, args, context) => {
-                return parseTokenFromHeaders(context)
-                    .then(verifyAndDecodeToken)
-                    .then(parseUserIdFromToken)
-                    .then(userId => {
-                        log.info('Adding watering schedule for: ', userId)
-                        return scheduleWateringFor(args.plants ?? {}, userId, args.timestamp, args.interval)
-                    })
-                    .catch(err => { log.error(err); throw new Error('Invalid Request') })
-            }
-        },
-        updateWateringSchedule: {
-            description: "Set up a reminder for when to water a plant next.",
-            type: WateringSchedule,
-            args: {
-                scheduleId: nonNullGqlString,
-                plants: {
-                    type: GraphQLList(plantInput)
-                },
-                timestamp: nonNullGqlString,
-                interval
-            },
-            resolve: async (_root, args, context) => {
-                return parseTokenFromHeaders(context)
-                    .then(verifyAndDecodeToken)
-                    .then(parseUserIdFromToken)
-                    .then(userId => {
-                        log.info('Updating schedule for: ', userId)
-                        return updateWateringSchedule(args.scheduleId, args.plants ?? {}, userId, args.timestamp, args.interval)
-                    })
-                    .catch(err => { log.error(err); throw new Error('Invalid Request') })
-            }
-        },
-        deleteWateringSchedule: {
-            description: "Remove a schedule",
-            type: Status,
-            args: {
-                id: nonNullGqlString,
-            },
-            resolve: async (_root, args, context) => {
-                return parseTokenFromHeaders(context)
-                    .then(verifyAndDecodeToken)
-                    .then(parseUserIdFromToken)
-                    .then(userId => {
-                        log.info('Removing schedule for: ', userId)
-                        return removeWateringScheduleById(args.id)
-                    } )
-                    .catch(err => { log.error(err); throw new Error('Invalid Request') })
-            }
-        }
+        nextWateringDateFor: nextWateringDateFor,
+        updateWateringSchedule: updateWateringScheduleForUser,
+        deleteWateringSchedule: deleteWateringSchedule
     })
 })
